(Abstract) com_camera, sys_camera
