+++ Abstract/world.ts
@@ -1,10 +1,13 @@
+import {Camera} from "./components/com_camera.js";
 import {Transform} from "./components/com_transform.js";
 
 const enum Component {
+    Camera,
     Transform,
 }
 
 export const enum Has {
+    Camera = 1 << Component.Camera,
     Transform = 1 << Component.Transform,
 }
 
@@ -13,5 +16,6 @@ export interface World {
     Signature: Array<number>;
 
     // Component data
+    Camera: Array<Camera>;
     Transform: Array<Transform>;
 }
